fix: modify regex pattern
I found a gap in the game table where it shouldn't be one. GTG for some time used a U+FE0F : VARIATION SELECTOR-16 [VS16] {emoji variation selector} along with a black square. This is added to the matching pattern. Some misses because string was shorter than 6 emojis. This requirement has been relaxed.