Feature/16/stop streak on gaps (#19)
* fix: modify regex pattern

I found a gap in the game table where it shouldn't be one. GTG for some time used a U+FE0F : VARIATION SELECTOR-16 [VS16] {emoji variation selector} along with a black square. This is added to the matching pattern. Some misses because string was shorter than 6 emojis. This requirement has been relaxed.

* feat: add gap termination.

Streaks starts counting on first result value and stops on a 0 or gap.

* feat: add gap termination in stats counting

* hotfix: fix error when using /stats in dms

Command checks if message is a dm befor trying getting guild member name.

* hotfix: fix streak counting error

Current streak was counted on the first streak, not, last.